Relocate `_TensorCodec` to `constant_op.py`.
In its current location, the codec induces a cyclic dependency between
`constant_op` and `tensor` by way of `ops.py`. Moving it to `constant_op`
breaks this dependency cycle.

Renames the codec to reflect its functionality and location.

PiperOrigin-RevId: 528849980
